requirements.txt
absl-py==0.11.0
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
decorator==4.4.2
Flask==1.1.2
Flask-Cors==3.0.10
google-auth==1.25.0
google-auth-oauthlib==0.4.2
grpcio==1.35.0
idna==2.10
imageio==2.9.0
importlib-metadata==3.4.0
itsdangerous==1.1.0
Jinja2==2.11.3
kiwisolver==1.3.1
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.4
networkx==2.5
numpy==1.20.0
oauthlib==3.1.0
opencv-python==4.5.1.48
pandas==1.2.1
Pillow==8.1.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
scikit-image==0.18.1
scipy==1.6.0
six==1.15.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tifffile==2021.2.1
torch==1.7.1
torchvision==0.8.2
typing-extensions==3.7.4.3
urllib3==1.26.3
Werkzeug==1.0.1
zipp==3.4.0
